+++ plugins/gs/src/components/DeploymentDetails/index.ts
@@ -1,0 +1,1 @@
+export { DeploymentDetails } from './DeploymentDetails';